% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.effect.size.coxPH.R
\name{get.effect.size.coxPH}
\alias{get.effect.size.coxPH}
\title{Effect size and odds ratio for SNP profiles}
\usage{
get.effect.size.coxPH(pheno = pheno, pheno.col, days.col = days,
  probs = probs, sdp.file = "~/Desktop/R/QTL/WD/HS_Sanger_SDPs.txt.bgz",
  markers, threshold = 5.05,
  dir = "/Users/elijah/Desktop/R/QTL/WD/2. Binomial Mapping/")
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}
}

