% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.effect.size.R
\name{get.effect.size}
\alias{get.effect.size}
\title{Effect size and odds ratio for SNP profiles}
\usage{
get.effect.size(pheno = pheno, pheno.col, probs = probs,
  sdp.file = "~/Desktop/R/QTL/WD/HS_Sanger_SDPs.txt.bgz", markers,
  threshold = 5.05,
  dir = "/Users/elijah/Desktop/R/QTL/WD/2. Binomial Mapping/")
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}
}

