% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sig.thr.R
\name{get.sig.thr}
\alias{get.sig.thr}
\title{Given a set of maximum LOD scores, calculate the significance thresholds
for the given levels. Note that we use the method of Broman et.al., Genetics,
2006 to calculate the X chromosoem threshold.
Daniel Gatti
dan.gatti@jax.org
Oct. 26, 2015
Arguments: perms: either a vector with the maximum LOD score from permutations
                  or, if Xchr == TRUE, a matrix with colnames 'A' and 'X'
                  containing the maximumum autosomal and X chr LOD scores from
                  permutations.
           alpha: numeric vector, one or more numerical value that are the
                  alpha levels of the significane thresholds.
           Xchr: logical indicating whether the X chromosome should have
                 a separate threshold. If TRUE, then perms must be a matrix
                 with 2 named columns as described above.
NOTE: If using p-values, don't forget to take -log10(p.value) of the
      permuations before calling this function.}
\usage{
get.sig.thr(perms, alpha = 0.05, Xchr = TRUE)
}
\description{
Given a set of maximum LOD scores, calculate the significance thresholds
for the given levels. Note that we use the method of Broman et.al., Genetics,
2006 to calculate the X chromosoem threshold.
Daniel Gatti
dan.gatti@jax.org
Oct. 26, 2015
Arguments: perms: either a vector with the maximum LOD score from permutations
                  or, if Xchr == TRUE, a matrix with colnames 'A' and 'X'
                  containing the maximumum autosomal and X chr LOD scores from
                  permutations.
           alpha: numeric vector, one or more numerical value that are the
                  alpha levels of the significane thresholds.
           Xchr: logical indicating whether the X chromosome should have
                 a separate threshold. If TRUE, then perms must be a matrix
                 with 2 named columns as described above.
NOTE: If using p-values, don't forget to take -log10(p.value) of the
      permuations before calling this function.
}

