% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSD.coxph.R
\name{GRSD.coxph}
\alias{GRSD.coxph}
\title{Genome wide association mapping using CoxPH}
\usage{
GRSD.coxph(pheno, pheno.col, days.col, probs, K, addcovar, markers, snp.file,
  outdir = "~/Desktop/files/", tx = c("Gamma", "HZE", "Unirradiated",
  "All"), sanger.dir = "~/Desktop/R/QTL/WD/HS.sanger.files/")
}
\arguments{
\item{pheno:}{data.frame containing phenotypes in columns and samples in
rows. Rownames must contain sample IDs.}

\item{pheno.col:}{the phenotype column to map. surv = Surv(pheno$cat.days, pheno$cataract)}

\item{probs:}{3D numeric array containing the haplotype probabilities
for each sample. Samples in rows, founders in columns,
markers in slices. Samnple IDs must be in rownames. Marker
names must be in dimnames(probs)[[3]].}

\item{K:}{List of kinship matrices, one per chromosome in markers.}

\item{addcovar:}{data.frame of additive covariates to use in the mapping.
Sample IDs must be in rownames.}

\item{markers:}{data.frame containing at least 3 columns with marker names,
chr, Mb postion.}

\item{snp.file:}{character string containing the full path to the Sanger
SNP file containing genomic positions and SNPs. This file
is created using condense.sanger.snps().}
}
\details{
Dec. 29, 2015
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}
}

