% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSD.binom.fast.R
\name{GRSDbinom.fast}
\alias{GRSDbinom.fast}
\title{Association mapping on autosomal chromosomes with binary variable outcomes optimized for speed by
pre-formating sanger snps to sdp files.

Performs association mapping in multiparent mouse populations.}
\usage{
GRSDbinom.fast(obj, pheno, pheno.col, addcovar, tx,
  sanger.dir = "~/Desktop/R/QTL/WD/HS.sanger.files/")
}

