% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDbinom.xchr.fast.R
\name{GRSDbinom.xchr.fast}
\alias{GRSDbinom.xchr.fast}
\title{Association mapping on the X chromosome with binary variable outcomes, optimized for speed by
pre-formating sanger snps to sdp files.

Performs association mapping in multiparent mouse populations.}
\usage{
GRSDbinom.xchr.fast(obj, pheno, pheno.col, addcovar, tx,
  sanger.dir = "~/Desktop/R/QTL/WD/HS.sanger.files/")
}

