% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDcoxph.perms.R
\name{GRSDcoxph.perms}
\alias{GRSDcoxph.perms}
\title{Permutation analysis for CoxPH model in HZE.}
\usage{
GRSDcoxph.perms(perms, chr = 1:19, Xchr = FALSE, pheno, pheno.col, days.col,
  probs, K, addcovar, markers, snp.file, outdir = "/home/ubuntu", tx = "",
  sanger.dir = "home/ubuntu/HS.sanger.files/")
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}
}

