% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDcoxph.R
\name{GRSDcoxph}
\alias{GRSDcoxph}
\title{Association mapping with survival outcomes using the CoxPH model.}
\usage{
GRSDcoxph(obj)
}
\description{
Performs association mapping in multiparent mouse populations.
}

