% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDcoxph.xchr.R
\name{GRSDcoxph.xchr}
\alias{GRSDcoxph.xchr}
\title{Association mapping with survival outcomes using the CoxPH model.}
\usage{
GRSDcoxph.xchr(obj)
}
\description{
Performs association mapping in multiparent mouse populations.
}

