% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDcoxph.xchr4perms.R
\name{GRSDcoxph.xchr4perms}
\alias{GRSDcoxph.xchr4perms}
\title{Association mapping with survival outcomes using the CoxPH model.}
\usage{
GRSDcoxph.xchr4perms(obj, pheno, pheno.col, surv, addcovar, tx, sanger.dir)
}
\description{
Association mapping with survival outcomes using the CoxPH model.
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}

Performs association mapping in multiparent mouse populations.
}

