% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDpoisson.xchr.R
\name{GRSDpoisson.xchr}
\alias{GRSDpoisson.xchr}
\title{Association mapping on the X chromosome with poisson dist.

Performs association mapping in multiparent mouse populations.}
\usage{
GRSDpoisson.xchr(obj, pheno, pheno.col, tx)
}

