% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Allow for multiple, superimposed QTL plotting in mice.
Borrowed gratefully from Kevin Ushey and modified slightly.}
\usage{
manhattan_plot(pval, bp, chr, groups = NULL, cutoff = NULL,
  xlab = "Chromosome (base-pair position)",
  ylab = expression(paste(-log[10](italic(p)))), transform = TRUE,
  cex = 0.5, ...)
}
\description{
Arguments:
}

