% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hs.R
\name{plot.hs}
\alias{plot.hs}
\title{Plot function (w/ binning to average markers and max LOD)}
\usage{
\method{plot}{hs}(qtl, bin.width = 1000, ...)
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}
}

