% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scanone.assoc.perms.R
\name{Scanone.assoc.perms}
\alias{Scanone.assoc.perms}
\title{Permutation analysis for continuous traits in HZE ion exposed genetically heterogeneous mice.}
\usage{
Scanone.assoc.perms(perms, chr = 1:19, Xchr = FALSE, pheno, pheno.col,
  probs, K, addcovar, markers, snp.file, outdir = "~/Desktop/files",
  tx = "")
}
\author{
Elijah F. Edmondson <elijah.edmondson@gmail.com>
}

