% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDassoc.perms.R
\name{GRSDassoc.perms}
\alias{GRSDassoc.perms}
\title{Permutation analysis for HZE.}
\usage{
GRSDassoc.perms(perms, chr = 1:19, Xchr = FALSE, pheno, pheno.col, probs, K,
  addcovar, markers, snp.file, outdir = "~/Dropbox/Rstudio.cloud/WD",
  tx = "", sanger.dir = "~/Desktop/R/QTL/WD/HS.sanger.files/")
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}
}

