% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDbinom.R
\name{GRSDbinom}
\alias{GRSDbinom}
\title{Association mapping on autosomal chromosomes with binary variable outcomes.

Performs association mapping in multiparent mouse populations.}
\usage{
GRSDbinom(obj, pheno, pheno.col, addcovar, tx)
}

