% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSDpoisson.xchr.R
\name{GRSDpoisson.xchr}
\alias{GRSDpoisson.xchr}
\title{Association mapping on the X chromosome with poisson dist.}
\usage{
GRSDpoisson.xchr(obj, pheno, pheno.col, tx)
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}

Performs association mapping in multiparent mouse populations.
}

