% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scanone.assoc.perms.R
\name{Scanone.assoc.perms}
\alias{Scanone.assoc.perms}
\title{Permutation analysis for continuous traits in HZE ion exposed genetically heterogeneous mice.}
\usage{
Scanone.assoc.perms(perms, pheno = pheno, pheno.col, probs, K = K,
  tx = "", addcovar = addcovar, markers = MM_snps, sdp.file = sdp.file,
  ncl = 4)
}
\author{
Elijah F Edmondson, \email{elijah.edmondson@gmail.com}
}

